+++ GcproExtensionApp/LL/EL/FormRollStandPhoenix.cs
@@ -1185,7 +1185,7 @@ namespace GcproExtensionApp
             if (ComboDiagram.SelectedItem != null)
             {
                 selectedDiagram = ComboDiagram.SelectedItem.ToString();
-                objMDDYZ.Diagram = MDDYZ.ParseInfoValue(selectedDiagram, AppGlobal.FIELDS_SEPARATOR, AppGlobal.NO_DIAGRAM);
+                objMDDYZ.Diagram = (int)MDDYZ.ParseInfoValue(selectedDiagram, AppGlobal.FIELDS_SEPARATOR, AppGlobal.NO_DIAGRAM);
             }
             ///<Page></Page>
             objMDDYZ.Page = txtPage.Text;
