+++ GcproExtensionApp/LL/MA/FormMA_Discharger.cs
@@ -1368,7 +1368,7 @@ namespace GcproExtensionApp
             if (ComboDiagram.SelectedItem != null)
             {
                 selectedDiagram = ComboDiagram.SelectedItem.ToString();
-                myDischarger.Diagram = Discharger.ParseInfoValue(selectedDiagram, AppGlobal.FIELDS_SEPARATOR, AppGlobal.NO_DIAGRAM);
+                myDischarger.Diagram = (int)Discharger.ParseInfoValue(selectedDiagram, AppGlobal.FIELDS_SEPARATOR, AppGlobal.NO_DIAGRAM);
             }
             ///<Page></Page>
             myDischarger.Page = txtPage.Text;
