+++ GcproExtensionApp/LL/MA/FormMA_MotorWithBypass.cs
@@ -1602,7 +1602,7 @@ namespace GcproExtensionApp
             if (ComboDiagram.SelectedItem != null)
             {
                 selectedDiagram = ComboDiagram.SelectedItem.ToString();
-                myMotorWithBypass.Diagram = MotorWithBypass.ParseInfoValue(selectedDiagram, AppGlobal.FIELDS_SEPARATOR, AppGlobal.NO_DIAGRAM);
+                myMotorWithBypass.Diagram = (int)MotorWithBypass.ParseInfoValue(selectedDiagram, AppGlobal.FIELDS_SEPARATOR, AppGlobal.NO_DIAGRAM);
             }
             ///<Page></Page>
             myMotorWithBypass.Page = txtPage.Text;
